[#169167752] Removed component and finished the new screen
Removed component and finished the new screen
