+++ ts/navigation/routes.ts
@@ -40,8 +40,10 @@ const ROUTES = {
   PAYMENT_PICK_PAYMENT_METHOD: "PAYMENT_PICK_PAYMENT_METHOD",
   PAYMENT_CONFIRM_PAYMENT_METHOD: "PAYMENT_CONFIRM_PAYMENT_METHOD",
   PAYMENT_PICK_PSP: "PAYMENT_PICK_PSP",
+  // CIE
   CIE_EXPIRED_SCREEN: "CIE_EXPIRED_SCREEN",
   CIE_VALID_SCREEN: "CIE_VALID_SCREEN",
+  CIE_SUCCESS_SCREEN: "CIE_SUCCESS_SCREEN",
 
   // Pin
   PIN_LOGIN_NAVIGATOR: "PIN_LOGIN_NAVIGATOR",
